refactor: Game.java 删除光标冗余代码并清理 (~360行)
- 删除 Game.java 中已迁移至 CursorServiceManager 的全部光标方法
  (initializeLocalCursorRenderers, destroyLocalCursorRenderers,
   syncCursorWithStream, startCursorService, stopCursorService,
   handleCursorUpdate, setSystemOrOverlayCursor, restoreDefaultCursor,
   updateCursorServiceState 及相关字段)
- 保留 refreshLocalCursorState 作为委托方法供外部调用
- 清理 11 个未使用的 import 和死常量 KEEP_ALIVE_CHANNEL_ID
- 修复 mayReduceRefreshRate() 调用为 DisplayModeManager 静态方法
- Android.mk: nanors/rs.c → rswrapper.c (正确的多ISA编译入口)
- 添加 nanors/deps/obl 和 nanors/deps include 路径
- build.gradle: 限制 ABI 为 arm64-v8a/armeabi-v7a (x86 不兼容 rswrapper)
